+++ src/pynwb/ndx_ophys_devices/__init__.py
@@ -34,7 +34,7 @@ DichroicMirrorModel = get_class("DichroicMirrorModel", "ndx-ophys-devices")
 OpticalFilterModel = get_class("OpticalFilterModel", "ndx-ophys-devices")
 BandOpticalFilterModel = get_class("BandOpticalFilterModel", "ndx-ophys-devices")
 EdgeOpticalFilterModel = get_class("EdgeOpticalFilterModel", "ndx-ophys-devices")
-OpticalLensModel = get_class("OpticalLensModel", "ndx-ophys-devices")
+ObjectiveLensModel = get_class("ObjectiveLensModel", "ndx-ophys-devices")
 
 # Device classes
 OpticalFiber = get_class("OpticalFiber", "ndx-ophys-devices")
@@ -45,7 +45,7 @@ DichroicMirror = get_class("DichroicMirror", "ndx-ophys-devices")
 OpticalFilter = get_class("OpticalFilter", "ndx-ophys-devices")
 BandOpticalFilter = get_class("BandOpticalFilter", "ndx-ophys-devices")
 EdgeOpticalFilter = get_class("EdgeOpticalFilter", "ndx-ophys-devices")
-OpticalLens = get_class("OpticalLens", "ndx-ophys-devices")
+ObjectiveLens = get_class("ObjectiveLens", "ndx-ophys-devices")
 __all__ = [
     "LensPositioning",
     "FiberInsertion",
@@ -58,7 +58,7 @@ __all__ = [
     "OpticalFilterModel",
     "BandOpticalFilterModel",
     "EdgeOpticalFilterModel",
-    "OpticalLensModel",
+    "ObjectiveLensModel",
     "OpticalFiber",
     "ExcitationSource",
     "PulsedExcitationSource",
@@ -67,6 +67,6 @@ __all__ = [
     "OpticalFilter",
     "BandOpticalFilter",
     "EdgeOpticalFilter",
-    "OpticalLens",
+    "ObjectiveLens",
 ]
 del load_namespaces, get_class