+++ src/pynwb/tests/test_constructors.py
@@ -19,8 +19,8 @@ from ndx_ophys_devices.testing import (
     mock_BandOpticalFilter,
     mock_EdgeOpticalFilterModel,
     mock_EdgeOpticalFilter,
-    mock_OpticalLensModel,
-    mock_OpticalLens,
+    mock_ObjectiveLensModel,
+    mock_ObjectiveLens,
     mock_ExcitationSourceModel,
     mock_ExcitationSource,
     mock_PulsedExcitationSource,
@@ -90,12 +90,12 @@ def test_constructor_edge_optical_filter():
     mock_EdgeOpticalFilter()
 
 
-def test_constructor_optical_lens_model():
-    mock_OpticalLensModel()
+def test_constructor_objective_lens_model():
+    mock_ObjectiveLensModel()
 
 
-def test_constructor_optical_lens():
-    mock_OpticalLens()
+def test_constructor_objective_lens():
+    mock_ObjectiveLens()
 
 
 def test_constructor_excitation_source_model():