+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## New Features
 - Added auto-publish.yml GitHub Action to automatically publish new versions to PyPI upon GitHub Release.
 
+## Deprecations and Changes
+  - Renamed ``OpticalLens`` to ``ObjectiveLens``
+
 # v0.3.1 (September 19, 2025)
 
 ### New Features